ET use compiler macros in ops
Use new compiler macros to fix operator library builds.
Additional issues,
isnan and signbit supported types are only on float, double and long double. Apply static cast to avoid msvc compiler error.

Differential Revision: [D78854136](https://our.internmc.facebook.com/intern/diff/D78854136/)

[ghstack-poisoned]

+++ kernels/portable/cpu/op_amax.cpp
@@ -55,7 +55,7 @@ Tensor& amax_out(
           for (const auto out_ix : c10::irange(begin, end)) {
             out_data[out_ix] = plan.execute<CTYPE>(
                 [](CTYPE v, CTYPE max_v) {
-                  return std::isnan(v) || v > max_v ? v : max_v;
+                  return std::isnan(static_cast<float>(v)) || v > max_v ? v : max_v;
                 },
                 out_ix);
           }
